Dojo: Restructure packages (#62)
- package level renamed to dojo
- package level_1, level_2, level_3, level_4 added
- package builder added

closes #61
