+++ dojo-dungeon/src/dojo/rooms/level_1/Room1.java
@@ -1,10 +1,13 @@
-package level.rooms;
+package dojo.rooms.level_1;
 
 import contrib.level.generator.graphBased.RoomGenerator;
 import core.level.utils.DesignLabel;
 import core.level.utils.LevelSize;
 import core.utils.components.path.IPath;
 import core.utils.components.path.SimpleIPath;
+import dojo.rooms.LevelRoom;
+import dojo.rooms.Room;
+import dojo.rooms.SearchRoom;
 
 public class Room1 extends SearchRoom {
   private static final int monsterCount = 5;
@@ -16,7 +19,7 @@ public class Room1 extends SearchRoom {
   private static final String keyDescription = "A key to unlock the next room.";
   private static final IPath keyTexture = new SimpleIPath("items/key/gold_key.png");
 
-  Room1(
+  public Room1(
       LevelRoom levelRoom,
       RoomGenerator gen,
       Room nextRoom,
