+++ dojo-dungeon/src/dojo/rooms/level_1/Room3.java
@@ -1,4 +1,4 @@
-package level.rooms;
+package dojo.rooms.level_1;
 
 import contrib.components.InteractionComponent;
 import contrib.hud.dialogs.OkDialog;
@@ -9,6 +9,8 @@ import core.components.PositionComponent;
 import core.level.utils.DesignLabel;
 import core.level.utils.LevelSize;
 import core.utils.components.path.SimpleIPath;
+import dojo.rooms.LevelRoom;
+import dojo.rooms.Room;
 import java.io.IOException;
 import java.util.HashSet;
 import java.util.Random;
@@ -58,7 +60,7 @@ public class Room3 extends Room {
 
   private int correctAnswerCount = 0;
 
-  Room3(
+  public Room3(
       LevelRoom levelRoom,
       RoomGenerator gen,
       Room nextRoom,
