+++ dojo-dungeon/src/dojo/rooms/level_2/Room4.java
@@ -1,10 +1,13 @@
-package level.rooms;
+package dojo.rooms.level_2;
 
 import contrib.level.generator.graphBased.RoomGenerator;
 import core.level.utils.DesignLabel;
 import core.level.utils.LevelSize;
 import core.utils.components.path.IPath;
 import core.utils.components.path.SimpleIPath;
+import dojo.rooms.LevelRoom;
+import dojo.rooms.Room;
+import dojo.rooms.SearchRoom;
 
 public class Room4 extends SearchRoom {
   private static final int monsterCount = 5;
@@ -17,7 +20,7 @@ public class Room4 extends SearchRoom {
   private static final String keyDescription = "This gem opens the door to the next room.";
   private static final IPath keyTexture = new SimpleIPath("items/resource/saphire.png");
 
-  Room4(
+  public Room4(
       LevelRoom levelRoom,
       RoomGenerator gen,
       Room nextRoom,
