+++ dojo-dungeon/src/dojo/rooms/level_2/Room5.java
@@ -1,4 +1,4 @@
-package level.rooms;
+package dojo.rooms.level_2;
 
 import contrib.components.InteractionComponent;
 import contrib.hud.dialogs.OkDialog;
@@ -10,15 +10,21 @@ import core.level.utils.DesignLabel;
 import core.level.utils.LevelSize;
 import core.utils.IVoidFunction;
 import core.utils.components.path.SimpleIPath;
+import dojo.compiler.DojoCompiler;
+import dojo.rooms.LevelRoom;
+import dojo.rooms.Room;
+import dojo.rooms.TaskRoom;
+import dojo.tasks.Task;
 import java.io.IOException;
 import java.util.Set;
-import level.tasks.Task;
 
-public class Room6 extends TaskRoom {
-  private final String FILENAME1 = "Room5ToRefactor.java";
-  private final String title = "Refactoring";
+public class Room5 extends TaskRoom {
+  private final String FILENAME1 = "../dojo-dungeon/todo-assets/r5/Monster.java";
+  private final String FILENAME2 = "../dojo-dungeon/todo-assets/r5/MyMonster.java";
+  private final String CLASS_NAME = "MyMonster";
+  private final String title = "Monster besiegen";
 
-  Room6(
+  public Room5(
       LevelRoom levelRoom,
       RoomGenerator gen,
       Room nextRoom,
@@ -39,22 +45,31 @@ public class Room6 extends TaskRoom {
     IVoidFunction empty = () -> {};
     addTask(
         new Task(
-            this,
-            "task1",
-            empty,
-            () ->
-                OkDialog.showOkDialog(
-                    String.format(
-                        "Refactor die Datei %s und lasse die korrigierte Datei von einem Tutor überprüfen.",
-                        FILENAME1),
-                    title,
-                    empty),
-            (t1) -> {
-              OkDialog.showOkDialog(
-                  "Das war schon alles, die Tür zum nächsten Level ist geöffnet.", title, empty);
-              return true;
-            },
-            empty));
+                this,
+                "task1",
+                empty,
+                () ->
+                    OkDialog.showOkDialog(
+                        String.format(
+                            "Implementiere die Datei %s, nach der Vorgabe in %s. Wenn das Monster besiegt ist, soll sich die Tür zum nächsten Raum öffnen.",
+                            FILENAME2, FILENAME1),
+                        title,
+                        empty),
+                (t1) -> {
+                  DojoCompiler.TestResult results =
+                      new DojoCompiler().spawnMonsterToOpenTheDoor(FILENAME2, CLASS_NAME, this);
+                  if (results.passed()) {
+                    OkDialog.showOkDialog(
+                        "Ok! " + results.messages(),
+                        title,
+                        () -> OkDialog.showOkDialog("Das Monster ist gespawnt!", title, empty));
+                    return true;
+                  }
+                  OkDialog.showOkDialog("Fehler: " + results.messages(), title, empty);
+                  return false;
+                },
+                empty)
+            .setShouldOpenDoors(false));
 
     // Create questioner
     Entity questioner = new Entity();
@@ -76,7 +91,7 @@ public class Room6 extends TaskRoom {
                         },
                         () ->
                             OkDialog.showOkDialog(
-                                "Die Aufgabe ist abgeschlossen!", title, empty))));
+                                "Das Monster ist bereits gespawnt!", title, empty))));
 
     // Add questioner to room
     addRoomEntities(Set.of(questioner));
